vtirt/requirements.txt
dotmap==1.3.30
numpy==1.24.1
pyro_ppl==1.8.3
scikit_learn==1.2.0
scipy==1.11.1
torch==1.13.1
tqdm==4.64.1
